update rclone to use conda not direct download, move tests outside of package
