+++ datashuttle/datashuttle.py
@@ -6,10 +6,10 @@ from typing import Union, cast
 
 import paramiko
 
-from datashuttle.datashuttle import configs
-from datashuttle.datashuttle.utils_mod import rclone_utils, utils
-from datashuttle.datashuttle.utils_mod.decorators import requires_ssh_configs
-from datashuttle.datashuttle.utils_mod.directory_class import Directory
+from datashuttle import configs
+from datashuttle.utils_mod import rclone_utils, utils
+from datashuttle.utils_mod.decorators import requires_ssh_configs
+from datashuttle.utils_mod.directory_class import Directory
 
 # --------------------------------------------------------------------------------------------------------------------
 # Project Manager Class
