+++ datashuttle/utils_mod/decorators.py
@@ -1,6 +1,6 @@
 from functools import wraps
 
-from .utils import raise_error
+from datashuttle.utils_mod.utils import raise_error
 
 
 def requires_ssh_configs(func):
