+++ tests/test_configs.py
@@ -3,10 +3,10 @@ import pathlib
 import warnings
 
 import pytest
+import test_utils
 import yaml
 
-from datashuttle.datashuttle.datashuttle import DataShuttle
-from datashuttle.datashuttle.tests import test_utils
+from datashuttle.datashuttle import DataShuttle
 
 TEST_PROJECT_NAME = "test_configs"
